refactor: remove hoc v2
