+++ apps/artemis-web/components/layout/layout.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 import dynamic from 'next/dynamic';
 import Head from 'next/head';
-import { useRouter } from 'next/router';
 
 export default class Layout extends React.Component<any> {
   render() {
-    // const user = this.props.jwt ? this.props.jwt.user : null;
     const { children } = this.props;
     // const Footer = dynamic(() => import('../footer/footer'));
     const Header = dynamic(() => import('../header/header'));
@@ -29,7 +27,7 @@ export default class Layout extends React.Component<any> {
           ></script>
         </Head>
         <div className="layout">
-          <Header jwt={this.props.jwt} loading={this.props.loading} />
+          <Header />
           {children}
           {/* <Footer /> */}
         </div>
