+++ apps/artemis-web/components/sign-in/sign-in.tsx
@@ -1,28 +1,25 @@
-import React, { useEffect } from 'react';
 import Button from '@material-ui/core/Button';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import TextField from '@material-ui/core/TextField';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Checkbox from '@material-ui/core/Checkbox';
-import Link from '@material-ui/core/Link';
-import Grid from '@material-ui/core/Grid';
-import Typography from '@material-ui/core/Typography';
-
 import {
-  makeStyles,
-  createMuiTheme,
-  ThemeProvider,
-} from '@material-ui/core/styles';
-import Container from '@material-ui/core/Container';
-import {
-  orange,
-  lightBlue,
   deepOrange,
   deepPurple,
+  lightBlue,
+  orange,
 } from '@material-ui/core/colors';
-
-import { useState } from 'react';
+import Container from '@material-ui/core/Container';
+import CssBaseline from '@material-ui/core/CssBaseline';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+import Grid from '@material-ui/core/Grid';
+import Link from '@material-ui/core/Link';
+import {
+  createMuiTheme,
+  makeStyles,
+  ThemeProvider,
+} from '@material-ui/core/styles';
+import TextField from '@material-ui/core/TextField';
+import Typography from '@material-ui/core/Typography';
 import { useRouter } from 'next/router';
+import React, { useState } from 'react';
 
 const palletType = 'dark';
 const darkState = false;
