+++ apps/artemis-web/lib/helpers.ts
@@ -1,5 +1,4 @@
 import jwt from 'jsonwebtoken';
-import { useFetch } from '../hooks/useJWT';
 
 export function extractUser(req) {
   if (!req.user) return null;
@@ -12,15 +11,6 @@ export function extractUser(req) {
     role,
     lastLogin,
   };
-}
-
-export function getUser() {
-  const { status, data } = useFetch('/api/jwt');
-  const jwt = data ? parseJwt(data) : null;
-  const user = jwt ? jwt.user : null;
-  const loading = status !== 'fetched';
-
-  return [user, loading];
 }
 
 export function parseJwt(token) {
