+++ apps/artemis-web/pages/_app.tsx
@@ -1,11 +1,8 @@
 import './styles.css';
-import React, { useEffect } from 'react';
+import React from 'react';
 import { ApolloProvider } from '@apollo/client';
 import { useApollo } from '../utils/graphql';
 import Layout from '../components/layout/layout';
-import { useJWT, useFetch } from '../hooks/useJWT';
-import { parseJwt } from '../lib/helpers';
-import { useRouter } from 'next/router';
 
 function MyApp({ Component, pageProps }) {
   const client = useApollo(
@@ -13,13 +10,11 @@ function MyApp({ Component, pageProps }) {
     process.env.GRAPHQL_URI,
     process.env.GRAPHQL_WS_URI
   );
-  const { status, data } = useFetch('/api/jwt');
-  const jwt = null; // data ? parseJwt(data) : null;
 
   return (
     <ApolloProvider client={client}>
-      <Layout jwt={jwt} loading={status !== 'fetched'}>
-        <Component jwt={jwt} loading={status !== 'fetched'} {...pageProps} />
+      <Layout>
+        <Component {...pageProps} />
       </Layout>
     </ApolloProvider>
   );
