+++ apps/artemis-web/pages/overview.tsx
@@ -1,21 +1,20 @@
 import Head from 'next/head';
-import React, { useEffect } from 'react';
+import { useRouter } from 'next/router';
+import React from 'react';
 import HijackTableComponent from '../components/ongoing-hijack-table/ongoing-hijack-table';
 import StatsTable from '../components/stats-table/stats-table';
-import withAuth from '../components/with-auth/with-auth';
 import { useGraphQl } from '../hooks/useGraphQL';
 import { useJWT } from '../hooks/useJWT';
-import { useRouter } from 'next/router';
-import { getUser } from '../lib/helpers';
 
 const OverviewPage = (props) => {
   if (process.env.NODE_ENV === 'development') {
     if (typeof window !== 'undefined') {
+      // eslint-disable-next-line @typescript-eslint/no-var-requires
       const { worker } = require('../mocks/browser');
       worker.start();
     }
   }
-  const [user, loading] = getUser();
+  const [user, loading] = useJWT();
 
   const router = useRouter();
   if (!user && !loading) router.push('signin');
