+++ apps/artemis-web/pages/signup.tsx
@@ -1,14 +1,21 @@
 import dynamic from 'next/dynamic';
 import Head from 'next/head';
-import withAuth from '../components/with-auth/with-auth';
+import { useRouter } from 'next/router';
 import React from 'react';
+import { useJWT } from '../hooks/useJWT';
 
 const SignupPage = (props) => {
   const SignUpComponent = dynamic(() =>
     import('../components/sign-up/sign-up')
   );
 
-  const user = props.user;
+  const [user, loading] = useJWT();
+  const router = useRouter();
+  if (user && !loading) {
+    console.log(user);
+    if (user.role === 'pending') router.push('pending');
+    else router.push('overview');
+  }
 
   return (
     <>
@@ -17,7 +24,7 @@ const SignupPage = (props) => {
         <title>ARTEMIS - Sign Up</title>
       </Head>
       <div id="page-container">
-        {!user && (
+        {!user && !loading && (
           <div id="content-wrap" style={{ paddingBottom: '5rem' }}>
             <div className="container d-flex align-items-center flex-column">
               <SignUpComponent />
@@ -29,4 +36,4 @@ const SignupPage = (props) => {
   );
 };
 
-export default withAuth(SignupPage, 'RIA');
+export default SignupPage;
